be/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
uuid==1.30
python-dotenv==1.0.0
alembic==1.12.1
opencv-python==4.8.1.78
numpy==1.26.1
pillow==10.1.0
cloudinary==1.36.0
ultralytics==8.3.112
torch>=2.0.0
websockets==11.0.3
imageio~=2.31.0
imageio-ffmpeg~=0.4.8
yt_dlp==2025.4.30
requests==2.31.0
pydantic[email]
pytz==2024.2
